+++ app/src/main/java/com/decoutkhanqindev/dexreader/presentation/screens/home/components/MangaBanner.kt
@@ -48,7 +48,6 @@ import com.decoutkhanqindev.dexreader.presentation.model.manga.MangaModel
 import com.decoutkhanqindev.dexreader.presentation.model.value.language.LanguageValue
 import com.decoutkhanqindev.dexreader.presentation.model.value.manga.MangaContentRatingValue
 import com.decoutkhanqindev.dexreader.presentation.model.value.manga.MangaStatusValue
-import com.decoutkhanqindev.dexreader.presentation.screens.common.animateItemOnAppear
 import com.decoutkhanqindev.dexreader.presentation.screens.common.badges.MangaGenreChip
 import com.decoutkhanqindev.dexreader.presentation.screens.common.badges.MangaRatingChip
 import com.decoutkhanqindev.dexreader.presentation.screens.common.badges.MangaStatusBadge
@@ -111,7 +110,6 @@ fun MangaBanner(
         }
         .fillMaxSize()
         .padding(horizontal = 8.dp)
-        .animateItemOnAppear()
         .onClick(MaterialTheme.shapes.medium) { onItemClick(manga.id) }
     ) {
       MangaCoverArt(