+++ app/src/main/java/com/decoutkhanqindev/dexreader/presentation/screens/profile/components/sections/ProfileHistoryItem.kt
@@ -18,7 +18,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.decoutkhanqindev.dexreader.presentation.model.user.ReadingHistoryModel
-import com.decoutkhanqindev.dexreader.presentation.screens.common.animateItemOnAppear
 import com.decoutkhanqindev.dexreader.presentation.screens.common.image.MangaCoverArt
 import com.decoutkhanqindev.dexreader.presentation.screens.common.onClick
 import com.decoutkhanqindev.dexreader.presentation.screens.common.shimmerLoading
@@ -48,7 +47,6 @@ fun ProfileHistoryItem(
 
   Card(
     modifier = modifier
-      .animateItemOnAppear()
       .onClick(shape = MaterialTheme.shapes.medium) { onSelected() },
     shape = MaterialTheme.shapes.medium,
     colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surfaceVariant),